Add Storage.java
